backend/requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.15.0
pydantic>=1.8
supabase
PyJWT
openai
